WebService/requirements.txt
aiofiles==24.1.0
alembic==1.14.1
asgiref==3.8.1
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.13.1
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dict2xml==1.7.6
Flask==3.0.0
Flask-Caching==2.3.0
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
ghp-import==2.1.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.50
mkdocs-material-extensions==1.3.1
mysql-connector-python==9.2.0
mysqlclient==2.2.7
orjson==3.10.15
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
pickleDB==0.9.2
platformdirs==4.3.6
pycryptodome==3.21.0
Pygments==2.19.1
pymdown-extensions==10.14.1
PyMySQL==0.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
xmltodict==0.14.2
zipp==3.21.0
